+++ src/WatchCompass.Application/Application/DependencyInjection/ApplicationServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
 using Microsoft.Extensions.DependencyInjection;
+using WatchCompass.Application.UseCases.DiscoverMovies;
 using WatchCompass.Application.UseCases.Recommendations;
 using WatchCompass.Application.UseCases.MovieDetails;
 using WatchCompass.Application.UseCases.SearchMovies;
@@ -13,6 +14,7 @@ public static class ApplicationServiceCollectionExtensions
     public static IServiceCollection AddApplication(this IServiceCollection services)
     {
         services.AddScoped<SearchMoviesUseCase>();
+        services.AddScoped<DiscoverMoviesByGenreUseCase>();
         services.AddScoped<GetRecommendationsUseCase>();
         services.AddScoped<GetMovieDetailsUseCase>();
         services.AddScoped<GetGenresUseCase>();